+++ R/colnames_accessors.R
@@ -71,19 +71,30 @@ get_colnames <- function(df) {
     get_epoch_colnames(df)
   } else {
     cli::cli_abort(c("x" = "Unknown data type: {.val {type}}.",
-                     "i" = "Data type must be \"sessions\" or \"epochs\"."))
+                     "i" = "Data type must be \"sessions\" or \"epochs\".",
+                     "i" = "Use {.fn set_data_type} to set the data type attribute."))
   }
 }
 
 #' Set column names for a data frame (sessions or epochs)
 #'
-#' This function sets the "col" attribute for a data frame with a named list of column names.
-#' @param df A data frame containing either sessions or epochs data
+#' This function updates the "col" attribute for a dataframe, overwriting existing column names with the ones provided.
+#' @param df A data frame
 #' @param col A named list of column names
 #' @returns The data frame with updated "col" attribute
 #' @family internal
 #' @export
+#' @examples
+#' sessions <- set_colnames(
+#'  example_sessions,
+#'  list(night = "calendar_date", sleep_period = "duration_sleep")
+#' )
 set_colnames <- function(df, col) {
-  attr(df, "col") <- col
+  existing_col <- attr(df, "col")
+  merged_col <- existing_col
+  for (nm in names(col)) {
+    merged_col[[nm]] <- col[[nm]]
+  }
+  attr(df, "col") <- merged_col
   df
 }