+++ R/compliance.R
@@ -11,9 +11,37 @@
 #' @examples
 #' filtered_sessions <- set_min_time_in_bed(example_sessions, 2)
 set_min_time_in_bed <- function(sessions, min_time_in_bed, return_mask = FALSE) {
+  check_session_colnames(sessions, "time_in_bed")
   col <- get_session_colnames(sessions)
 
   mask <- sessions[[col$time_in_bed]] >= min_time_in_bed * 60 * 60
+
+  mask[is.na(mask)] <- FALSE
+
+  if (return_mask) {
+    mask
+  } else {
+    sessions[mask, ]
+  }
+}
+
+#' Set minimum sleep period
+#'
+#' @param sessions The sessions dataframe
+#' @param min_sleep_period The minimum sleep period in hours
+#' @param return_mask If TRUE, return a logical vector indicating which sessions meet the minimum sleep period requirement
+#' @details This function uses columns:
+#' - `sleep_period`
+#' @returns The sessions dataframe with only the sessions that meet the minimum sleep period requirement, or a logical vector if `return_mask` is TRUE
+#' @export
+#' @family filtering
+#' @examples
+#' filtered_sessions <- set_min_sleep_period(example_sessions, 2)
+set_min_sleep_period <- function(sessions, min_sleep_period, return_mask = FALSE) {
+  check_session_colnames(sessions, "sleep_period")
+  col <- get_session_colnames(sessions)
+
+  mask <- sessions[[col$sleep_period]] >= min_sleep_period * 60 * 60
 
   mask[is.na(mask)] <- FALSE
 
@@ -39,6 +67,7 @@ set_min_time_in_bed <- function(sessions, min_time_in_bed, return_mask = FALSE) 
 #' @examples
 #' filtered_sessions <- set_session_start_time_range(example_sessions, "22:00", "06:00")
 set_session_start_time_range <- function(sessions, from_time, to_time, return_mask = FALSE) {
+  check_session_colnames(sessions, "session_start")
   col <- get_session_colnames(sessions)
 
   session_times <- update_date(sessions[[col$session_start]], "0000-01-01")
@@ -76,6 +105,7 @@ set_session_start_time_range <- function(sessions, from_time, to_time, return_ma
 #' @examples
 #' filtered_sessions <- set_session_sleep_onset_range(example_sessions, "22:00", "06:00")
 set_session_sleep_onset_range <- function(sessions, from_time, to_time, return_mask = FALSE) {
+  check_session_colnames(sessions, "time_at_sleep_onset")
   col <- get_session_colnames(sessions)
 
   session_times <- update_date(sessions[[col$time_at_sleep_onset]], "0000-01-01")
@@ -109,6 +139,7 @@ set_session_sleep_onset_range <- function(sessions, from_time, to_time, return_m
 #' @examples
 #' filtered_sessions <- remove_sessions_no_sleep(example_sessions)
 remove_sessions_no_sleep <- function(sessions, return_mask = FALSE) {
+  check_session_colnames(sessions, "sleep_period")
   col <- get_session_colnames(sessions)
 
   mask <- sessions[[col$sleep_period]] > 0
@@ -133,6 +164,7 @@ remove_sessions_no_sleep <- function(sessions, return_mask = FALSE) {
 #' @examples
 #' duplicate_sessions <- get_non_complying_sessions(example_sessions)
 get_non_complying_sessions <- function(sessions) {
+  check_session_colnames(sessions, "night")
   col <- get_session_colnames(sessions)
   dup_mask <- duplicated(sessions[, c(col$night, col$subject_id)]) |
     duplicated(sessions[, c(col$night, col$subject_id)], fromLast = TRUE)
@@ -153,6 +185,8 @@ get_non_complying_sessions <- function(sessions) {
 #' filtered_sessions <- set_session_start_time_range(example_sessions, "22:00", "06:00")
 #' removed_sessions <- get_removed_sessions(example_sessions, filtered_sessions)
 get_removed_sessions <- function(sessions, filtered_sessions) {
+  check_session_colnames(sessions, c("id", "sleep_period"))
+  check_session_colnames(filtered_sessions, c("id", "sleep_period"))
   col <- get_session_colnames(sessions)
 
   if (nrow(filtered_sessions) > nrow(sessions)) {