+++ R/filtering.R
@@ -4,10 +4,9 @@
 #' @param sessions The sessions dataframe
 #' @param return_mask If TRUE, returns a logical vector indicating which epochs belong to the specified sessions
 #' @details This function uses sessions columns:
-#' - `session_start`
-#' - `session_end`
+#' - `id`
 #' And epoch columns:
-#' - `timestamp`
+#' - `session_id`
 #' @returns The epochs dataframe with only the epochs that belong to the specified sessions, or a logical vector if `return_mask` is TRUE
 #' @export
 #' @examples
@@ -18,6 +17,8 @@
 #' @seealso [filter_by_night_range()] to filter sessions by night range.
 #' @family filtering
 filter_epochs_from_sessions <- function(epochs, sessions, return_mask = FALSE) {
+  check_session_colnames(sessions, "id")
+  check_epoch_colnames(epochs, "session_id")
   scol <- get_session_colnames(sessions)
   ecol <- get_epoch_colnames(epochs)
 
@@ -37,7 +38,7 @@ filter_epochs_from_sessions <- function(epochs, sessions, return_mask = FALSE) {
 
   mask <- epochs[[ecol$session_id]] %in% sessions[[scol$id]]
 
-  # If there are unmatched session_ids, set display to FALSE
+  # If there are unmatched session_ids, set mask to FALSE
   mask[is.na(mask)] <- FALSE
 
   if (return_mask) {
@@ -62,6 +63,7 @@ filter_epochs_from_sessions <- function(epochs, sessions, return_mask = FALSE) {
 #' @examples
 #' filtered_sessions <- filter_by_night_range(example_sessions, "2025-04-07", "2025-04-10")
 filter_by_night_range <- function(sessions, from_night, to_night, return_mask = FALSE) {
+  check_session_colnames(sessions, "night")
   col <- get_session_colnames(sessions)
 
   from_night <- if (is.null(from_night)) min(sessions[[col$night]]) else from_night
@@ -91,6 +93,7 @@ filter_by_night_range <- function(sessions, from_night, to_night, return_mask = 
 #' @param return_mask If TRUE, returns a logical vector indicating which sessions belong to subjects within the specified age range
 #' @details This function uses columns:
 #' - `birth_year`
+#' - `night`
 #' @returns The sessions dataframe with only the sessions that belong to subjects within the specified age range,
 #' or a logical vector if `return_mask` is TRUE
 #' @importFrom rlang .data
@@ -99,21 +102,18 @@ filter_by_night_range <- function(sessions, from_night, to_night, return_mask = 
 #' @examples
 #' filtered_sessions <- filter_by_age_range(example_sessions_v1, min_age = 11, max_age = 18)
 filter_by_age_range <- function(sessions, min_age = NULL, max_age = NULL, return_mask = FALSE) {
+  check_session_colnames(sessions, c("birth_year", "night"))
   col <- get_session_colnames(sessions)
 
-  if (is.null(col$birth_year)) {
-    cli::cli_abort(c("!" = "The sessions table does not contain a birth year column."))
-  }
-
-  min_age <- if (is.null(min_age)) min(lubridate::year(sessions[[col$session_start]]) - sessions[[col$birth_year]]) else min_age
-  max_age <- if (is.null(max_age)) max(lubridate::year(sessions[[col$session_start]]) - sessions[[col$birth_year]]) else max_age
+  min_age <- if (is.null(min_age)) min(lubridate::year(sessions[[col$night]]) - sessions[[col$birth_year]]) else min_age
+  max_age <- if (is.null(max_age)) max(lubridate::year(sessions[[col$night]]) - sessions[[col$birth_year]]) else max_age
 
   if (min_age > max_age) {
     cli::cli_abort(c("!" = "min_age must be before max_age."))
   }
 
-  mask <- (sessions[[col$birth_year]] >= (lubridate::year(sessions[[col$session_start]]) - max_age) &
-             sessions[[col$birth_year]] <= (lubridate::year(sessions[[col$session_start]]) - min_age))
+  mask <- (sessions[[col$birth_year]] >= (lubridate::year(sessions[[col$night]]) - max_age) &
+             sessions[[col$birth_year]] <= (lubridate::year(sessions[[col$night]]) - min_age))
 
   mask[is.na(mask)] <- FALSE
 
@@ -138,11 +138,8 @@ filter_by_age_range <- function(sessions, min_age = NULL, max_age = NULL, return
 #' @examples
 #' filtered_sessions <- filter_by_sex(example_sessions_v1, "M")
 filter_by_sex <- function(sessions, sex, return_mask = FALSE) {
+  check_session_colnames(sessions, "sex")
   col <- get_session_colnames(sessions)
-
-  if (is.null(col$sex)) {
-    cli::cli_abort(c("!" = "The sessions table must contain a Sex column."))
-  }
 
   mask <- sessions[[col$sex]] %in% sex
 
@@ -169,6 +166,7 @@ filter_by_sex <- function(sessions, sex, return_mask = FALSE) {
 #' @examples
 #' filtered_sessions <- select_subjects(example_sessions, c("sub_01JNDH3Z5NP0PSV82NFBGPV31X"))
 select_subjects <- function(sessions, subject_ids, return_mask = FALSE) {
+  check_session_colnames(sessions, "subject_id")
   col <- get_session_colnames(sessions)
 
   if (sum(sessions[[col$subject_id]] %in% subject_ids) == 0) {
@@ -202,6 +200,7 @@ select_subjects <- function(sessions, subject_ids, return_mask = FALSE) {
 #' @examples
 #' filtered_sessions <- select_devices(example_sessions, c("VTGVSRTHCA"))
 select_devices <- function(sessions, device_ids, return_mask = FALSE) {
+  check_session_colnames(sessions, "device_id")
   col <- get_session_colnames(sessions)
 
   if (sum(sessions[[col$device_id]] %in% device_ids) == 0) {