+++ R/sleep_bubbles.R
@@ -6,23 +6,31 @@
 #' - `sleep_period`
 #' - `night`
 #' @param color_by The variable to color the bubbles by. Can be "default" or any other column name in the sessions dataframe.
+#' @param bubble_size The size of the bubbles. Default is 10.
 #' @returns A ggplot object containing the sleep bubbles graph.
 #' @importFrom rlang .data
 #' @export
 #' @family plot sessions
-plot_sleep_bubbles <- function(sessions, color_by = "default") {
+plot_sleep_bubbles <- function(sessions, color_by = "default", bubble_size = 10) {
+  check_session_colnames(sessions, c("sleep_period", "night"))
   col <- get_session_colnames(sessions)
 
   sessions <- sessions |>
-    dplyr::filter(!is.na(.data[[col$time_at_sleep_onset]]) & !is.na(.data[[col$time_at_wakeup]])) |>
+    remove_sessions_no_sleep() |>
     dplyr::mutate(sleep_duration = .data[[col$sleep_period]] / 3600)
 
   if (color_by != "default" && color_by %in% names(sessions)) {
-    sessions$color_group <- as.factor(sessions[[color_by]])
-    color_levels <- levels(sessions$color_group)
-    color_map <- stats::setNames(scales::hue_pal()(length(color_levels)), color_levels)
-    color_aes <- ggplot2::aes(x = .data[[col$night]], y = .data$sleep_duration, color = .data$color_group)
-    color_scale <- ggplot2::scale_color_manual(values = color_map, name = color_by)
+    color_var <- sessions[[color_by]]
+    if (is_iso8601_datetime(color_var)) {
+      color_var <- parse_time(color_var) |> update_date(date = "1970-01-01")
+      color_aes <- ggplot2::aes(x = .data[[col$night]], y = .data$sleep_duration, color = color_var)
+    } else if (is.numeric(color_var)) {
+      color_aes <- ggplot2::aes(x = .data[[col$night]], y = .data$sleep_duration, color = color_var)
+    } else {
+      sessions$color_group <- as.factor(color_var)
+      color_aes <- ggplot2::aes(x = .data[[col$night]], y = .data$sleep_duration, color = .data$color_group)
+    }
+    color_scale <- get_color_scale(color_var)
   } else {
     sessions$color <- suppressWarnings(dplyr::case_when(
       sessions$sleep_duration >= 6 & sessions$sleep_duration <= 9 ~ scales::col_numeric(
@@ -42,13 +50,13 @@ plot_sleep_bubbles <- function(sessions, color_by = "default") {
       ymin = 6, ymax = 9,
       fill = "lightgrey", alpha = 0.5
     ) +
-    ggplot2::geom_point(size = 20, alpha = 0.5) +
+    ggplot2::geom_point(size = bubble_size, alpha = 0.5) +
     color_scale +
     ggplot2::labs(
       x = NULL,
       y = "Sleep Duration (hours)",
       title = NULL,
-      color = if (color_by != "default" && color_by %in% names(sessions)) color_by else NULL
+      color = NULL
     ) +
     ggplot2::theme_minimal(base_size = 16) +
     ggplot2::theme(